Refactor the contents of the batch details modal into a modal container and child form, so we can easily implement the exact same form in the inventory create view later. Also added validation for the quantities to ensure that at least one of the quantities has a non-zero value
