+++ src/strings/csv/en.csv
@@ -345,6 +345,7 @@ ENTER_SIX_DIGIT_KEY,Enter 6-digit key,
 ENTER_SUBSET_WEIGHT_AND_COUNT,"Enter Subset Weight and Count ratio to determine approximate seed count when withdrawing by weight from a count quantity. (Conversely, the ratio will determine approximate remaining quantity when withdrawing by count from a weight quantity.) These fields are useful when counting tiny seeds is not practical and only required if you intend to withdraw from this accession to a nursery.",
 ENVIRONMENTAL_NOTES,Environmental Notes,
 ENVIRONMENTAL_NOTES_PLACEHOLDER,"Important notes about landscape, climate, and environmental conditions",
+ERROR_BATCH_SEED_COUNT,Transfers must include at least 1 seed,
 ERROR_LOAD_SPECIES,Unable to load species for the organization,
 ERROR_LOAD_SUB_LOCATIONS,Unable to load sub locations for facility,
 EST_READY_DATE,Est. Ready Date,
